+++ src/Color/ColorFileOld.cpp
@@ -44,7 +44,7 @@ namespace lce::color {
         return io.getData();
     }
 
-    uint64_t ColorFileOld::getSize() const {
+    size_t ColorFileOld::getSize() const {
         uint32_t size = 8; // 4 for version 4 for count
         for (const auto &color : colors) {
             size += 2;
